Reorganize `no_std` example
